Extend run.py with a --examiner option for direct exam of fragment files.
Add the notion of a CallbackExcavation which is intended to be programmed
at runtime. Start to expose examiners via a module which can be referenced
by name and combine these to allow run.py to program the examination of a
single specified file.

Note that a mechanism is added to allow the artifact type checks performed
by examiners to be separable from their construction, but examiners must
be converted to expose this class method. Do this as a first for WangText.